Fix failing tests.
Using CREATE OR REPLACE VIEW instead of CREATE only because all the
tables and views are created twice in the test_session fixture. That is
because of the use of DeferredBase and Base, which share the same
metadata. This seems wrong to me, but prefer to change only one thing
at a time.