Create monthly avg of temp extrema as views, with tests.
These are views, not materialized views yet because mat view helpers don't work yet.

Also not sure if I like adding DaysInMonth function here as well as in crmp repo, but it seems to be the only way to have it available in the ORM (for testing). Ideally it would be defined in one place only (here).